requirements.txt
django
sqlalchemy
psycopg2-binary
pytest
alembic
python-dotenv
requests
rich>=13.0.0
